backend/requirements.txt
# Standard packages from PyPI (install these first)
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pytesseract>=0.3.10
azure-ai-contentsafety>=1.0.0
azure-core>=1.28.0
azure-identity>=1.13.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# EasyOCR and its dependencies
easyocr>=1.7.0
opencv-python-headless>=4.5.0
scipy>=1.7.0
scikit-image>=0.18.0
python-bidi>=0.4.0
Pillow>=9.0.0

# Dashboard dependencies
pandas>=2.1.4
openpyxl>=3.1.2
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.21.0

# Additional scientific computing
matplotlib>=3.5.0
seaborn>=0.11.0
imageio>=2.9.0

# PyTorch packages (will be installed from PyTorch CPU index)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.3.0
torchvision>=0.18.0